DRL_for_Controlling_Cartpole/Cartpole_code/requirements.txt
gym==0.18.0
pyyaml==5.4.1
matplotlib==3.3.4